libmount: fix merge of options in fsinfo_buf2fs()
The mount attribute MOUNT_ATTR_RELATIME is defined as 0 in the fsinfo()
header files.

The vfs option relatime is included in the vfs options proc string only
if noatime and strictatime are not set so this option gets ommitted when
using fsinfo().

This maybe should be fixed in mnt_optstr_apply_flags() but it's a special
case that may cause side effects for non fsinfo() use. Or perhaps fsinfo()
should set the flag to its MS_* counterpart similar to the other MOUNT_ATTR_*
flags, not sure?

In any case I needed this set so added this hack.

Signed-off-by: Ian Kent <raven@themaw.net>