libmount: update fsinfo includes
After a make in the kernel source tree all that's needed to build against
a kernel with fsinfo() is:
CFLAGS="-I <kernel source base>/usr/include"

Also, using the above include path will define __NR_fsinfo so checking
if it's defined and seting it to -1 is enough to handle kernels without
fsinfo().

And move the include directives to the top of fs.c.

Signed-off-by: Ian Kent <raven@themaw.net>